+++ documentation/sphinx/source/release-notes.rst
@@ -28,6 +28,7 @@ Features
 * Added support for restoring multiple key ranges in a single restore job. `(PR #1190) <https://github.com/apple/foundationdb/pull/1190>`_
 * Depracated transaction option ``TRANSACTION_LOGGING_ENABLE``. Added two new transaction options ``DEBUG_TRANSACTION_IDENTIFIER`` and ``LOG_TRANSACTION`` that sets an identifier for the transaction and logs the transaction to the trace file respectively. `(PR #1200) <https://github.com/apple/foundationdb/pull/1200>`_
 * Clients can now specify default transaction timeouts and retry limits to all transactions through a database option. `(Issue #775) <https://github.com/apple/foundationdb/issues/775>`_
+* The "timeout" transaction option is no longer reset when the transaction is reset after a call to ``onError`` (as of API version 610). `(Issue #775) <https://github.com/apple/foundationdb/issues/775>`_
 
 Performance
 -----------
@@ -47,6 +48,7 @@ Status
 Bindings
 --------
 
+* API version updated to 610.
 * The API to create a database has been simplified across the bindings. All changes are backward compatible with previous API versions, with one exception in Java noted below. `(PR #942) <https://github.com/apple/foundationdb/pull/942>`_
 * C: ``FDBCluster`` objects and related methods (``fdb_create_cluster``, ``fdb_cluster_create_database``, ``fdb_cluster_set_option``, ``fdb_cluster_destroy``, ``fdb_future_get_cluster``) have been removed. `(PR #942) <https://github.com/apple/foundationdb/pull/942>`_
 * C: Added ``fdb_create_database`` that creates a new ``FDBDatabase`` object synchronously and removed ``fdb_future_get_database``. `(PR #942) <https://github.com/apple/foundationdb/pull/942>`_
@@ -73,7 +75,7 @@ Bindings
 Other Changes
 -------------
 
-* Migrated  to Boost 1.67. `(PR #1242) https://github.com/apple/foundationdb/pull/1242`_
+* Migrated  to Boost 1.67. `(PR #1242) <https://github.com/apple/foundationdb/pull/1242>`_
 
 Earlier release notes
 ---------------------